Add project to the issue.
Turns out that gh issues do not have project expanded but they have the
project URL. This add such info to be available to the issue in listings
and filters.

This also introduces a cache for requests it is intended to be used in
more places than the caching of projects.